minor read me updates and code strucure
