+++ src/yaml_default_config.yaml
@@ -15,8 +15,8 @@ model:
 quantum:
   backend: "local"  # "local" or "scaleway"    
   #use_scaleway: false     # False = use local simulation; True = use remote QPU via Scaleway
-  n_modes: 12
-  n_photons: 6
+  n_modes: 24  #142
+  n_photons: 10  #27
   postselect: 3
   n_samples: 1000  #
   circuit_type: "pdf"   # Options: "triangular", "rectangular", "pdf", "base"  # Options: "triangular", "rectangular"
@@ -27,7 +27,7 @@ quantum:
   local_backend: "SLOS"   # sim:sampling:l4  SLOS -- Perceval backend for local simulation (e.g., SLOS, CliffordClifford2017, MPS)
   scaleway_platform: "sim:sampling:2l4"  # sim:sampling:2l4  sim:sampling:p100 Remote platform (if use_scaleway=true)
   downsample_method: "pca"   # pca or "bilinear" (default)
-
+  #sim:sampling:2l4 platform can simulate up to 27 photons and 142 modes,
 
 training:
   epochs: 5
@@ -41,7 +41,7 @@ optimization:
   max_evals: 10
   space:
     n_modes: [6, 12, 24, 30]
-    n_photons: [2, 3, 6, 8]
+    n_photons: [6, 10]
     learning_rate: [0.0001, 0.001, 0.01]
